Amend send bill run statuses (#632)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-225

As part of our drive to standardise bill run statuses, we amend the send bill run process as follows:

At the start of the process, the bill run status is set to pending to show that the details are being updated and therefore subject to change;
Once this is done, the status is set to sending and the process continues to the transaction file creation stage.
After the file is created, the status is set to either billed or billing_not_required (as is currently the case).


* Implement `sending` status

We update `SendBillRunReferenceService` so it initially sets the bill run status to `pending`. Then, once the bill run updates have taken place, we set the status to `sending`.

* Update service comments with changes

* `SendTransactionFileService` now expects `sending`

`SendTransactionFileService` starts by validating the bill run to ensure that its status is `pending`. However we now use the status `sending`, so we update the service to expect this status instead.