CMEA-232 Add new V3 delete invoice endpoint
https://eaflood.atlassian.net/browse/CMEA-232

Since there is no change in behaviour between v2 and v3 we simply create a v3 route `/v3/{regimeSlug}/bill-runs/{billRunId}/invoices/{invoiceId}` and point it to the existing controller.

N.B. We are creating a series of `/v3` routes even where no change is required to give a clear distinction between a workflow that supports both SROC and PRESROC (v3) and one that only supports PRESROC (v2).