CMEA-234 Delete SRoC Licence (#635)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-234

We found that there was no need to update DeleteLicenceService to use the correct deminimis limit for the ruleset of the parent bill run; the work we had previously done to implement billRun.$deminimisLimit() was sufficient. Therefore, we simply create a v3 route DELETE /v3/{regimeSlug}/bill-runs/{billRunId}/licences/{licenceId} and point it to the existing controller.


* Create v3 route

Since we are not implementing a breaking change, we create our new v3 route and point it at the existing service, which we will be updating to support sroc alongside presroc

* Update unit test to cover sroc deminimis

We are already using the bill run's deminimis limit (based on its ruleset) to determine whether the deminimis flag should be set, so no additional changes are needed to make this sroc-compliant. We create new sroc-specific unit tests to confirm that this is the case.