Add additional presroc validation (#640)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-214

When we were implementing sroc charge validation, we realised that we could validate some fields in a better way than we were validating them for presroc. This change therefore updates the transaction translators so that this improved validation is also applied to the presroc ruleset, by moving the validation to the base class:
- We validate that if twoPartTariff is true then section127Agreement is also true
- We validate that twoPartTariff and compensationCharge are not both set to true
- We validate that the max length of chargePeriod is 150 chars. Note that this change is made to the transaction translator, rather than the charge translator.

For sroc fields where text has to be precisely capitalised to be accepted by the Rules Service, we performed a case-insensitive match against a list of possible values, which gives the exact capitalisation as well as allowing us to reject invalid input early instead of passing it to the Rules Service for it to reject. This is an improvement over the presroc validation, which simply capitalises each word. We therefore implement this for presroc (refactoring so that it is done by a function on the base class) and expand it to cover all such fields:
- loss
- source
- season
- eiucSource
- regionalChargingArea


* Move `twoPartTariff/section127Agreement` to base

If `twoPartTariff` is `true` then we want to ensure that `section127Agreement` is also `true`. We therefore move this rule from the sroc charge translator to the base charge translator so that it applies to both sroc and presroc.

* Move `compensationCharge/twoPartTariff` to base

We want to ensure that `twoPartTariff` and `compensationCharge` are not both set to `true`. We therefore move this rule from the sroc charge translator to the base charge translator so that it applies to both sroc and presroc.

* Move `chargePeriod` to base

We want to validate that the max length of `chargePeriod` is 150 chars. We therefore move this rule from the sroc transaction translator to the base transaction translator so that it applies to both sroc and presroc.

* Improved `loss` validation

We currently validate sroc `loss` against an array of possible values. We mirror this approach for presroc by moving the rule to the base translator, and having the presroc and sroc translators provide their own `_validLosses()`.

* Improved `source` validation

We update the presroc-only `source` validation to use a list of supported sources, instead of simply capitalising the provided value

* Improved `season` validation

We update the validation for `season` to validate against a list of possibilities instead of just capitalising the supplied text. We also remove the function used to capitalise text as we no need it.

* Improved `eiucSource` validation

We update the validation for `eiucSource` to validate against a list of possibilities.

* Improved `regionalChargingArea` validation

We update the validation for `regionalChargingArea` to validate against a list of possibilities.

* Add unit test and fix `eiucSource` validation

We add in a unit test to check that the improved string validation returns the correct capitalisation. This catches a small issue with the `eiucSource` validation, which we therefore fix.

* Refactor string validation

We use the `Joi.string().valid(...).insensitive()` validation in a number of places across both sroc and presroc translators. We therefore create a function for this on the base translator and update the individual ruleset translators to use it.