Align presroc and sroc `regionalChargingArea` validation (#644)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-214

Previously, presroc and sroc had a different list of valid values for `regionalChargingArea`. However we have had confirmation that these lists can be the same; therefore, we delete the lists from the individual translators, and put the single list we want to use on the base translator, so they both validate against the same list.

We also move the validation rule to the base validator, as we have agreed that the behaviour should be the same for both. This means a slight change in how the field is validated for presroc to bring it in line with sroc: previously, the field was always required for presroc, whereas now it is only required if `compensationCharge` is `true`. In practice, `regionalChargingArea` has always been sent through to us anyway but this change makes it clear in the code what the expected behaviour is.

* Move `_validRegionalChargingAreas()` to base

* Move `regionalChargingArea` validation to base