Remove `twoPartTariff` validation from presroc translator
We previously refactored our presroc and sroc charge validation so that validation we want to apply to both rulesets is placed in a shared base translator. We did this with our sroc validation which ensures that both `twoPartTariff` and `compensationCharge` aren't both true, as we wanted this to also apply to presroc. However, we overlooked removing the existing `twoPartTariff` validation from the presroc translator, which meant that the specific conditional validation was being overwritten with the basic "ensure this field is present" validation we had for presroc. We therefore remove this from the presroc translator to ensure that the conditional validation applies to both rulesets.