CMEA-220 Add SROC transaction number sequence
https://eaflood.atlassian.net/browse/CMEA-220

A transaction reference is assigned to each invocie to be billed when a bill run is `sent`. There is a need for the SROC references to be distinguishable from PRE-SROC. In the [TCM](https://github.com/DEFRA/sroc-tcm-admin) this is done by adding a `T` to the end of the reference.

We've been asked to do the same in the CM. We also need to maintain separate sequences.

So, this change handles updating the sequence counters to allow for a new SROC transaction number sequence.\