Update and add tests ahead of changes
SROC has a different format to PRESROC so we need to account for this in the tests. Ruleset becomes a critical param needed to make the service work so all calls need to be updated. Plus we add a test to cover when an unknown ruleset is specified (though this should never happen).