Refactor response
In line with the work done on `NextTransactionReferenceService`, we move from having a "one size fits all" response template that always tries to include a suffix (which would be blank for presroc) to having separate presroc and sroc templates, which are effectively the same but sroc has the `t` suffix.

We do this by moving the response out of `BaseNextFileReferenceService` and into `NextTransactionFileReferenceService`. We therefore also need a response in `NextCustomerFileReferenceService`. Finally, this refactoring means we no longe need `_fileFixedChar()`, so we can remove it from each service.