Create SRoC transaction file references (#649)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-221

Currently, presroc transaction files have a filename in the format `nalri50001`. We want sroc transaction files to have a filename in the format `nalri50001t`, ie. with a `t` suffix. However, we want the sroc and presroc sequences to be continuous, ie. `nalri50001t` would be followed by either `nalri50002` or `nalri50002t`, which would be followed by either `nalri50003` or `nalri50003t`, etc.

We implement this as follows:

- We update `BaseNextFileReferenceService` to no longer have a generic filename template in `_response()`, but instead expect services that extend it to provide their own `_response()` method;
- We add a `_response()` method to `NextTransactionFileReferenceService` which checks whether the ruleset is sroc or presroc and returns the appropriate filename;
- We also add a `_response()` method to `NextCustomerFileReferenceService` to return the expected filename format for customer files.


* Determine file suffix based on ruleset

* Additional unit test

* Fix comment typo

* Change `BaseNextFileReferenceService` to `this`

We don't need to explicitly state that we're calling methods on `BaseNextFileReferenceService`; we can simply call them on `this`

* Refactor response

In line with the work done on `NextTransactionReferenceService`, we move from having a "one size fits all" response template that always tries to include a suffix (which would be blank for presroc) to having separate presroc and sroc templates, which are effectively the same but sroc has the `t` suffix.

We do this by moving the response out of `BaseNextFileReferenceService` and into `NextTransactionFileReferenceService`. We therefore also need a response in `NextCustomerFileReferenceService`. Finally, this refactoring means we no longe need `_fileFixedChar()`, so we can remove it from each service.

* Rework unit tests

* Remove unneeded item from `rules_service.config.js`

Prior to refactoring, we put the filename suffix in `rules_service.config.js`. After refactoring we now explictly state the file format and therefore no longer need this, however we hadn't removed it.