Change en-dash in source names to regular hyphen (#652)
When reviewing sroc validation we noticed that some supported source names had an en-dash symbol in them, eg. Nene – Northampton, possibly as a result of copying the list from a Sharepoint document (which is prone to "helpfully" formatting such things). This is an issue as although it is almost indistinguishable from a regular hyphen - (ie. the normal dash key on a keyboard), it wouldn't be treated as the same when comparing strings. So if the CM received the string Nene - Northampton (with a regular dash) it would reject it.

We therefore standardise on a regular hyphen for these names, and update CalculateChargeSrocTranslator accordingly.