Make `regionalChargingArea` required for presroc (#657)
We previously changed the charge validation for the field `regionalChargingArea` as part of our work to bring presroc validation in line with sroc, by making the field optional if `compensationCharge` is set to `false`. This was done on the basis that this isn't part of the charge calculation for either ruleset, and therefore doesn't need to be sent to the Rules Service.

However, it turns out that the Rules Service still requires `regionalChargingArea` in order to calculate the `sucFactor` of a presroc charge, and therefore presroc requests will fail if this field isn't present. Note that `sucFactor` is not part of sroc, so is still optional for that ruleset.

We therefore revert our previous change and make `regionalChargingArea` a required field for presroc, regardless of `compensationCharge`.