Remove sroc deminimis limit (#663)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-256

We disable deminimis for sroc by setting the deminimis limit to `0`. We also update unit tests to demonstrate that even the smallest invoice is not identified as deminimis for sroc.


* Set sroc deminimis limit to 0

We disable deminimis for sroc by setting the limit to 0. We also update unit tests to demonstrate that even the smallest invoice is not identified as deminimis for sroc.

* Reorganise `DeleteLicenceService` unit tests

We split the sroc and presroc deminimis tests into their own `describe` blocks