Remove duff view bill run transaction endpoint
When doing a final comparison of v2 versus v3 routes in our [OpenAPI docs](https://github.com/DEFRA/sroc-charging-module-api-docs) (to check we hadn't missed any) @StuAA78 spotted that we had completely overlooked `GET /v2/{regimeSlug}/bill-runs/{billRunId}/transactions/{transactionId}`.

😮 😕 🤔

Hold on though. Way back when rebuilding the original `v1` API we had been handed there was the ability to view a transaction. So, we [Add stand-in view bill run transaction endpoint](https://github.com/DEFRA/sroc-charging-module-api/pull/211) with the intent to add the proper logic later.

Only we never did. The current dev team were able to demonstrate and argue that the WRLS team never used this endpoint or intended to. The CM Team it seemed were unaware that WRLS treat the data they hold as `main` and only seek the charge information we generate. There was an assumption they were always querying us when transaction information was needed which we proved was incorrect.

So, this change removes the duff endpoint so we don't confuse ourselves again in the future!