Amend presenter to handle sroc file references
We previously took the last 5 characters of the file reference to give us the value required by the transaction file header, eg `50003`. However for sroc transaction files, we want the last 6 characters, eg. `50003T`. Since we know that file references are always in the format `nalwi50003` or `nalwi50003T`, we can simply change this to use the 6th char of the file reference onwards.