Amend transaction file head presenter for sroc (#669)
https://eaflood.atlassian.net/browse/CMEA-257

As per AC1 in the card, the format of the transaction file head is pretty much the same for both presroc and sroc. However the way it turns a file reference into a file number needs to be changed.

The existing code takes a reference in the format `nalwi50003` and extracts the last 5 characters, ie. `50003`. However for sroc references like `nalwi50003t` we want to extract the last 6 characters and make it upper case, ie. `50003T`.

Since file references always have 5 letters at the start, we simply change this to extract the 6th character onwards and upper-case it.


* Amend presenter to handle sroc file references

We previously took the last 5 characters of the file reference to give us the value required by the transaction file header, eg `50003`. However for sroc transaction files, we want the last 6 characters, eg. `50003T`. Since we know that file references are always in the format `nalwi50003` or `nalwi50003T`, we can simply change this to use the 6th char of the file reference onwards.

* Fix failing `GeneratePresrocTransactionFileService` unit test