Add `prorataDays`/`lineAttr3` to sroc charge (#671)
https://eaflood.atlassian.net/browse/CMEA-257

We need to put the prorata days string in the sroc transaction file (as we do with presroc transaction files), however we overlooked this when setting up our sroc charge. Therefore we need to update `CalculateChargeSrocTranslator` so that it adds this in. We do this by refactoring the prorata days methods into the `CalculateChargeBaseTranslator`.