Format the exported transaction file for SROC
https://eaflood.atlassian.net/browse/CMEA-257

When an SROC bill run is `/send` we generate a transaction file. But the format is specific to presroc. The invoices that get generated by SOP using this information are different to the those used for SROC bill runs. So, to ensure SOP gets the information it needs we need to reformat the transaction export to match the requirements for SROC.