Get ready to test generator selection
`SendTransactionFileService` is now going to need to pick the appropriate GenerateTransactionFileService depending on the ruleset of the bill run.

We get the tests ready for this by renaming the existing stub (to make it clearer). We don't think there is a need to repeat _all_ the tests for each ruleset. But we do need a dedicated unit tests for the selector.

And though new bill runs created using the `/v3` endpoint will have `sroc` as the default ruleset currently our helper method creates them as `presroc` unless you express otherwise. This is why we have left all the tests using the presroc generate stub.