Generate sroc transaction file service (#670)
https://eaflood.atlassian.net/browse/CMEA-257

When an SROC bill run is `/send` we generate a transaction file. But the format is specific to presroc. The invoices that get generated by SOP using this information are different to those used for SROC bill runs. So, to ensure SOP gets the information it needs we need to reformat the transaction export to match the requirements for SROC.

In this change, we create `GenerateSrocTransactionFileService`, which uses `TransactionFileSrocBodyPresenter` (along with the regular head and tail presenters) to format a transaction file for sroc. We also update `SendTransactionFileService` to include logic to determine which file generation service to use based on the bill run's ruleset.

## Notes

We uncovered a small defect with `TransactionFileSrocBodyPresenter`. We have logic that checks whether `headerAttr2` and `regimeValue19` equal `1`. But these are stored as strings in the db so we fix them to check against `'1'` instead.

We also found that the tests for the SROC presenter were expecting the wrong values for columns 28 and 29. So, those tests have been fixed.