Store `abatementFactor` in `regimeValue11` (#680)
https://eaflood.atlassian.net/browse/CMEA-258

We need to find somewhere in the database to store `adjustmentFactor` as part of the above card. As part of our previous work of storing sroc-only data in unused fields, we ended up using all the existing generic fields to do so. This means that we potentially need to create a new db field for `adjustmentFactor` -- which we would like to avoid if possible so we don't need to run migrations on all our environments (including prod). Thankfully, a solution presents itself!

While looking over the existing data, we realised that the presroc `section126Factor` value which we currently store in the db as `regimeValue11` and the sroc `abatementFactor` value which we currently store as `regimeValue19` are simply two names for pretty much the same thing. (To be precise: the section 126 factor is an example of an abatement factor; we are simply giving it a more general name under sroc). This was in part identified by the fact that we pass both of these to the rules service under the same name `abatementAdjustment` (and we acknowledge that this name is a little confusing for an **abatement** factor, given that an **adjustment** factor is an entirely separate thing!).

Since `section126Factor` and `abatementFactor` are effectively the presroc and sroc names for the same thing, we therefore change our existing code so we now store `abatementFactor` as `regimeValue11` to match how we store `section126Factor`. This therefore frees up `regimeValue19` for us to store `adjustmentFactor`.


* Store `abatementFactor` in `regimeValue11`

* Add clarifying comment

We note in `RulesServiceSrocPresenter` where field names differ to their name elsewhere. We previously overlooked such a comment for `abatementAdjustment`, so we add one now.