Use `adjustmentFactor` in sroc transaction file (#683)
https://eaflood.atlassian.net/browse/CMEA-258

If `adjustmentFactor` is anything other than `1` then we want to add `Other` to the list of reductions in the sroc transaction file. We do this by amending `GenerateSrocTransactionFileService` to retrieve `regimeValue19` as part of the select query to get all the required data for the transaction file, then updating `TransactionFileSrocBodyPresenter` to add `Other` to the list of reductions based on its value.