Fix date formatting for node 16.14.0 (#685)
While working on a feature, we found that our unit tests were failing during CI but not locally, and the same test was passing yesterday in CI. The failing test was being caused by `toLocaleString()` now returning `Sept` instead of `Sep` for September, which breaks date formatting for transaction and customer files.

On investigation we found that the version of Node being used by Docker was previously 16.13.2 but was now 16.4.0. It appears that the underlying internationalisation data that node uses has changed between these versions. As stated in a [previous related error](https://github.com/nodejs/node/issues/38897):

> The purpose of the locale strings has always been to attempt to output something that will be readable to people with their own language and cultural preferences. Think of a general user viewing some financial information. As I mentioned, if your documents have a specific requirement, they fall somewhat outside of this.

Since the output of `toLocaleString()` cannot be relied on to be consistent, we ditch it completely and instead construct the date string ourselves.