Additional validation for `waterUndertaker` (#684)
https://eaflood.atlassian.net/browse/CMEA-255

We currently validate for charges and transactions that if `compensationCharge` is `true` then `waterUndertaker` is required. For sroc, we additionally want to validate that if `waterCompanyCharge` is `true` then `waterUndertaker` must be `true` (but not vice-versa, ie. if `waterUndertaker` is `true` then `waterCompanyCharge` could potentially be `false`).

We make this change by first moving our existing validation ("if `compensationCharge` is `true` then `waterUndertaker` is required") from `CalculateChargeBaseTranslator` to `CalculateChargePresrocTranslator`, as we can only replace rules in the specific translators -- we cannot extend them. We can then add our more specific validation to `CalculateChargeSrocTranslator`.

Note that we have confirmed that we only need to do this validation if `compensationCharge` is `true`; if it's `false` then we don't care whether or not `waterUndertaker` has been provided (even though it's likely we will always be given it), and if we _have_ received `waterUndertaker` in this scenario then we don't care about its value either.


* Move existing validation into presroc translator

We currently have validation in the base charge translator to ensure that `waterUndertaker` is a boolean. Since our additional validation will only apply to sroc, we therefore move the existing validation out of the base translator and into the presroc translator.

* Add additional validation to sroc translator