Make sroc factors mandatory (#686)
https://eaflood.atlassian.net/browse/CMEA-260

As per the card, we are making `abatementFactor` and `aggregateProportion` mandatory. We therefore add `.required()` to their validation definitions and remove the default values. We also bring them in line with the `adjustmentFactor` validation by enforcing that the value cannot be `0` or less.


* Create failing unit tests

We start by creating unit tests for checking the validation of `abatementFactor` and `aggregateProportion` which we know will fail

* Update validation so unit tests pass

* Remove `.only` from test suite