Fix how `got` is imported
We use the workaround of using `await import('got')` to import `got`. Since JS doesn't support top level await, we do this within an asyncronous function. For `RequestRulesServiceChargeService`, this is done right before we use it, in the method that sends the request to the Rules Service. For the unit tests, we do this within an IIFE as we need `got` to be available to all tests