Update `got` to v12 (#693)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-270

As of v12, the `got` dependency that we use to send requests to the Rules Service only supports ECMAScript Modules (ESM) and not Common JavaScript modules (CJS). This is a problem for us as the JS ecosystem is still very much in flux, and we cannot yet move across to ESM as not all of our dependencies support this.

To ensure we aren't stuck with a dependency which gradually becomes further and further out of date, we either replace it or find a workaround that allow us to continue using it. [A workaround that is officially recommended](https://github.com/sindresorhus/got/issues/1789) is to replace the usual CJS method of loading the dependency:
```js
require('got')
```
with this:
```js
await import('got')
```
The downside of this is that JS doesn't support `await` at the top level. We therefore need to do this within a function.

For `RequestRulesServiceChargeService`, this means importing it in the `_callRulesService` method (which is already asynchronous), immediately prior to using it.

In the unit tests, we need `got` to be defined at the top level as we attach a spy to it and monitor it within each unit test. We therefore import it in an asynchronous IIFE.

We also need to amend how we set up the request's timeout value, as this has changed in v12. [As stated in the docs](https://github.com/sindresorhus/got/blob/main/documentation/6-timeout.md), we now need to provide an object for `timeout` instead of a value, and we use `{ request: ... }` as this is the global timeout which matches the previous v11 behaviour.


* Update `got` to v12

* Fix how `got` is imported

We use the workaround of using `await import('got')` to import `got`. Since JS doesn't support top level await, we do this within an asyncronous function. For `RequestRulesServiceChargeService`, this is done right before we use it, in the method that sends the request to the Rules Service. For the unit tests, we do this within an IIFE as we need `got` to be available to all tests