Tidy unit tests
We initially created one unit test for each sort criteria, and a fourth unit test to cover the order the sorting is applied in. Since this last test effectively covers all eventualities, we delete the tests that separately check the order. We also make some other small changes to tidy up the remaining tests.