Finish up tasks
`app/tasks/index.js` contained some logic to check whether it was being used to require a file, or being called via `npm run` to kick off a task. Since we no longer use it to require files, we start by removing the `if` check so it always runs a task. We also remove the `modules.export` section. We then rename it to `npm_task_runner.js` to better reflect its purpose, and update `package.json` accordingly.