Remove `index.js` files (#698)
When we built the service, we used the convention of creating an `index.js` file in each main folder which `require`s each file of that type, and we would then include the files elsewhere by destructuring from `index.js`. For example, the file `services/index.js` would have a `require` for each service, and we would then use services elsewhere with:
```js
const { ExampleService } = require('../services')
```

As we continued to build the service, we encountered some issues with this approach. One is that we could not be consistent with this approach; a service cannot destructure from `services/index.js` because `services/index.js` already refers to that service, creating a circular dependency. So in a service, we have to require other services like this:
```js
const ExampleService = require('./example.service')
```

Another is that when the time comes to move the service across from Common JavaScript (which uses `require` to include other files) to EcmaScript Modules (which uses `import` to include other files), we cannot use destructuring in this way; we have to individually `import` each file. `import` also needs the precise filename including the `.js` extension, whereas `require` will take a guess that the file has this extension.

So, partly to be consistent across our code and partly in preparation for migrating to ESM, we make the following changes:
- we change all our `require` statements so they directly import files instead of destructuring from `index.js`;
- we delete the now-unneeded `index.js` files;
- we point to the precise filename with `.js` extension;
- we rename `db/index.js` to `db/db.js`, as it contains code rather than being a simple `index.js` file like the others;
- we rename `app/tasks/index.js` to `app/tasks/npm_task_runner.js` as it is also called from `package.json` when running tasks via `npm run`; and we remove the now-unnecessary logic to determine if it is being used to require tasks from within our code.


* Update controllers

We update all controllers to specifically require files instead of destructuring from `index.js`

* Update libs

Note that as our intent is to directly require our own files, we leave the destructuring of `Notifier` from `@airbrake/node`

* Update plugins

* Update routes

* Update services

* Update tasks

* Update translators

Note that most of the changes here are just to align with the standard of "dependencies first"

* Update models

The changes here are simply to align with our "dependencies first" standard plus keeping a line between each "type" of require (ie. imports, models etc.)

* Update `server.js`

* Update db

Note that we leave `db/index.js` alone -- we are eliminating the pattern of using an `index.js` which pulls in other files -- in this case, `db/index.js` contains the actual code we refer to, so we don't make any changes

* Prepare test files

We are going to use global search and replace to amend the test files. To do this we need to turn multiple destructures into one line per require, eg:
```
const { LicenceModel, TransactionModel } = require('../../../app/models')
```
becomes:
```
const { LicenceModel } = require('../../../app/models')
const { TransactionModel } = require('../../../app/models')
```

* Fix all test helper references

* Fix all model references

* Fix all service references

* Fix all lib references

* Fix all presenter references

* Fix all translator references

* Fix all config references

* Fix all db references

* Fix task references

* Remove unneeded comment

We spotted a comment that referred to us requiring a file directly instead of destructuring from `index.js`. Since we are moving to directly requiring all files, this is no longer relevant

* Fix a couple of references we missed earlier

* Delete `index.js` files

We delete the majority of `index.js` files, with the exception of `db/index.js` (which we will shortly rename) and the files under `test/support/fixtures`, which we still need to decide what we're doing with

* Rename `db/index.js` to `db/db.js`

Since we are moving away from using `index.js`, we rename the file `db/index.js` to `db/db.js` to ensure we are consistent

* Replace `app/tasks/index.js`

We realised that `app/tasks/index.js` contains some functionality that we may need to retain, so we replace it until we decide how to resolve this.

* Explicitly require `.js` files

CJS `require` does not need the `.js` extension to be a the end of the filename. However as preparation for later changing across to ESM `import`, which does need the exact filename, we add `.js` to the end of each filename that we require.

* Finish up tasks

`app/tasks/index.js` contained some logic to check whether it was being used to require a file, or being called via `npm run` to kick off a task. Since we no longer use it to require files, we start by removing the `if` check so it always runs a task. We also remove the `modules.export` section. We then rename it to `npm_task_runner.js` to better reflect its purpose, and update `package.json` accordingly.

* Fix linting error

Our linter throws an error for `npm_task_runner.js` as we are defining constants which are not referred to within the file. They do need to be defined though as they are used when calling it via `npm start`. We therefore disable the linting rule for these lines.

* Point routes to `.js` extension

We overlooked changing route files to use the `.js` extension

* Destructure `temporaryFilePath`

We noticed that we are able to destructure `temporaryFilePath` from `server.config.js` in a few places