Create `ValidateInvoiceService`
We need to change how the service is called -- instead of passing in an invoice and bill run id, we need to pass in the actual invoice and bill run objects. Rather than changing the service and updating its callers in one go, we instead start by creating the final `ValidateInvoiceService`. We will then migrate the individual callers across to using it.