Update "view invoice" to use new service
As with "delete bill run", we delete the test for a non-existent invoice. We also delete the test for an invoice not linked to the bill run, as this behaviour is part of `ValidateInvoiceService` and is tested there. And since we don't need `ValidateInvoiceService` to do anything, we can simply stub it out.