Change `FetchAndValidateInvoiceService` to `ValidateInvoiceService` (#699)
https://eaflood.atlassian.net/browse/CMEA-154

When we initially developed `FetchAndValidateInvoiceService`, we used it to fetch an invoice and perform a couple of checks on it:
1. that the invoice exists;
2. that the invoice is linked to the specified bill run.

We later created a plugin that automatically fetches an invoice when an invoice-related endpoint is used, and performs the first check. We now revisit our initial service to change it to be simply `ValidateInvoiceService`, which performs just the second check.


* Create `ValidateInvoiceService`

We need to change how the service is called -- instead of passing in an invoice and bill run id, we need to pass in the actual invoice and bill run objects. Rather than changing the service and updating its callers in one go, we instead start by creating the final `ValidateInvoiceService`. We will then migrate the individual callers across to using it.

* Update "delete bill run" to use new service

Note that we can delete the test for a non-existant invoice as we no longer handle this within the validation service -- it will already have been handled by `RequestInvoicePlugin` prior to reaching the controller

* Update "view invoice" to use new service

As with "delete bill run", we delete the test for a non-existent invoice. We also delete the test for an invoice not linked to the bill run, as this behaviour is part of `ValidateInvoiceService` and is tested there. And since we don't need `ValidateInvoiceService` to do anything, we can simply stub it out.

* Update comments based on review

* Delete `FetchAndValidateInvoiceService`

We overlooked the fact that we can delete `FetchAndValidateInvoiceService` now that we have completely replaced it with `ValidateInvoiceservice`!