Implement fix
We had overlooked updating the code in the controller to _not_ depend on what the validation service returns (because it doesn't return anything).

This meant we were validating that the invoice could be deleted and returning a `2xx` response but then erroring because we were passing in `undefined` to actually be deleted.