Extract factors with more than 1 dp
Based on what we'd seen in the ACs and in the water abstraction validation for incoming data, we made the assumption that a factor will have no more than 1 decimal place. However when we started investigating some failing tests, we realised that this assumption was incorrect, and that values such as `0.833` could potentially be returned by the Rules Service. We therefore update `_extractFactor()` to allow factors with more than 1 decimal place.